+++ lib/tilt/_slim.rb
@@ -1,0 +1,2 @@
+warn "Lazy loading of slim templates is deprecated and will be removed in Tilt 3.  Require slim manually.", uplevel: 1
+require "slim"